Refactor tests into one directory
