+++ src/tests/calculate.test.js
@@ -1,4 +1,4 @@
-import calculate from './calculate';
+import calculate from '../logic/calculate';
 
 const object = {
   total: null,
